+++ src/Elders.Cronus.AtomicAction.Redis/RELEASE_NOTES.md
@@ -1,3 +1,18 @@
+#### 0.2.0 - 10.11.2016
+* Update Cronus to 3.0.2
+* Expose additional settings
+* Fixed NullReferenceException when there is no value for key in redis
+
+#### 0.2.0-beta0003 - 10.11.2016
+* Update Cronus to 3.0.2
+
+#### 0.2.0-beta0002 - 08.11.2016
+* Expose additional settings
+* Fixed NullReferenceException when there is no value for key in redis
+
+#### 0.2.0-beta0001 - 18.03.2016
+* Cronus.AtomicAction.Redis with latest Cronus reference
+
 #### 0.1.1 - 18.03.2016
 * Updated RedLock reference
 * Added Example project